메시지(notes) 기능 베타에서 제거
베타 범위 축소·관리 부담 감소를 위해 커플/가족용 메시지(음성/텍스트 노트) 기능을 완전 제거. NativeTab.Messages 탭·라우트, VoiceMessagePanel·노트 UI, unreadMessageCount·messagesLocked, 노트 ViewModel 액션(sendNote/sendTtsNote/downloadNoteAudio/markNoteRead/refreshNotes 등)·NotesApi·노트 폴링(RemoteAlarmSyncWorker)·노트 알림 제거, 백엔드 /notes 마운트 해제(notes.ts는 휴면 유지로 재도입 용이). 공유 헬퍼(bearerOrMessage 등 5개)는 MainViewModelAuthActions로 이동 보존. People·가족·음성공유·결제·알람은 무손상.
