+++ apps/android-native/app/src/main/java/com/alarmtalk/app/network/AlarmTalkApi.kt
@@ -8,5 +8,4 @@ interface AlarmTalkApi :
     FamilyApi,
     CodeApi,
     BillingApi,
-    NotesApi,
     HolidayApi
