+++ apps/android-native/app/src/main/java/com/alarmtalk/app/ui/app/AlarmTalkAppHelpers.kt
@@ -82,7 +82,6 @@ internal val NativeTab.route: String
         NativeTab.Voices -> "voices"
         NativeTab.Alarms -> "alarms"
         NativeTab.People -> "people"
-        NativeTab.Messages -> "messages"
         NativeTab.Billing -> "billing"
         NativeTab.Menu -> "menu"
     }
