+++ apps/android-native/app/src/main/java/com/alarmtalk/app/ui/navigation/NavigationModels.kt
@@ -33,7 +33,6 @@ internal enum class NativeTab {
     Voices,
     Alarms,
     People,
-    Messages,
     Billing,
     Menu,
 }
